fixes tests

+++ routes/users.js
@@ -1,16 +1,23 @@
 import buildFormObj from '../lib/formObjectBuilder';
+import { isExist } from '../lib/tools';
 
 export default (router, { User }) => {
   router
     .get('users#list', '/users', async (ctx) => {
-      const users = await User.findAll();
-      ctx.render('users', { users });
+      // console.log(ctx.state.isSignedIn());
+      if (!ctx.state.isSignedIn()) {
+        ctx.flash.set('Please Log In For Access to This page');
+        ctx.redirect(router.url('sessions#new'));
+      } else {
+        const users = await User.findAll();
+        ctx.render('users/index', { users });
+      }
     })
     .get('users#new', '/users/new', (ctx) => {
       const user = User.build();
       ctx.render('users/new', { f: buildFormObj(user) });
     })
-    .post('users#create', '/users/new', async (ctx) => {
+    .post('users#create', '/users', async (ctx) => {
       const { form } = ctx.request.body;
       const user = User.build(form);
       try {
@@ -24,15 +31,29 @@ export default (router, { User }) => {
     .get('users#profile', '/users/:id/profile', async (ctx) => {
       const id = Number(ctx.params.id);
       const user = await User.findByPk(id);
-      // console.log(user);
-      ctx.render('users/profile', { user });
+      if (!isExist(user)) {
+        ctx.render('errors/index');
+      } else if (ctx.state.isSignedIn()) {
+        ctx.render('users/profile', { user });
+      } else {
+        ctx.flash.set('Please Log In For Access to This page');
+        ctx.redirect(router.url('sessions#new'));
+      }
     })
     .get('users#edit', '/users/:id/edit', async (ctx) => {
-      const { id } = ctx.params;
-      // console.log(id);
-      const user = await User.findByPk(Number(id));
-      // console.log(user);
-      ctx.render('users/edit', { f: buildFormObj(user), id });
+      if (ctx.state.isSignedIn()) {
+        const { id } = ctx.params;
+        if (Number(id) !== ctx.state.signedId()) {
+          ctx.flash.set('Profile Edit Does Not Allowed');
+          ctx.redirect(router.url('users#profile', id));
+        } else {
+          const user = await User.findByPk(Number(id));
+          ctx.render('users/edit', { f: buildFormObj(user), id });
+        }
+      } else {
+        ctx.flash.set('Please Log In For Access to This page');
+        ctx.redirect(router.url('sessions#new'));
+      }
     })
     .patch('users#update', '/users/:id', async (ctx) => {
       const { id } = ctx.params;
@@ -47,12 +68,24 @@ export default (router, { User }) => {
       }
     })
     .delete('users#delete', '/users/:id', async (ctx) => {
-      const userId = Number(ctx.params.id);
-      User.destroy({
-        where: { id: userId },
-      });
-      ctx.flash.set('You account was been deleted.');
-      ctx.session = {};
-      ctx.redirect(router.url('root'));
+      if (ctx.state.isSignedIn()) {
+        console.log(ctx.state.signedId());
+        const userId = Number(ctx.params.id);
+        console.log(userId);
+        if (ctx.state.isSignedIn() && ctx.state.signedId() === userId) {
+          User.destroy({
+            where: { userId },
+          });
+          ctx.flash.set('You account was been deleted.');
+          ctx.session = {};
+          ctx.redirect(router.url('root'));
+        } else {
+          ctx.flash.set('Profile Delete Does Not Allowed');
+          ctx.redirect(router.url('users#profile', userId));
+        }
+      } else {
+        ctx.render('errors/index');
+        ctx.status = 404;
+      }
     });
 };
